Use _POSIX_C_SOURCE, drop _BSD_SOURCE, _GNU_SOURCE (#5, #12)
The only need for _BSD_SOURCE is a single use of the trivial, but
non-standard, strdup(). The only need for _GNU_SOURCE is for getline().
This function was standardized by POSIX 10 years ago, so you only need
to ask for it with _POSIX_C_SOURCE.

Also added time.h which is only included by luck from the removed
feature test macros.

This is better than PR #5 because _DEFAULT_SOURCE isn't needed at all.