Fix out of bounds read of render buffer
When characters in a render buffer are compared
to the strings in the keywords array this is
possible that p + klen points out of the buffer
and memcmp results to out-of-bounds access.

To see that this is true just recompile kilo
using GCC with the option -fsanitize=address
and open any file using the fresh build of kilo.
Address Sanitizer points to the problem
immediately.