Avoid invalid memory access if rest of line is shorter than some keyword
The test for a keyword would try to read the entire length of each keyword
from the line position even if the rest of the line was too short to
contain that keyword.
A PR was already submitted for this but the author closed it before
it could be merged.