Fix display and edit of long lines
The accounting of text and cursor positions failed when editing text at the RH end of a long line, especially when `TAB` characters are in the LH part of the line. A function `getRenderCol()` is introduced that calculates the actual display width of the LH part of a row, and is used instead of the character width as needed.