remove python multiline comments
since python multiline string/comments start and end with the same
sequence, kilo is unable to figure out which is which once one removes
either one of the start or end markers of the section.
dealing with this would require a quite more elaborate engine than
currently implement, or doing a new syntax scan over the entire file.