don't redraw entire screen if we move cursor within visible area
for a cursor move that doesnt result in any scrolling, we now simply update
cursor position (and col/row in status bar) to remove the nasty flickering
caused by kilo on xterm when redrawing the entire screen everytime a cursor
is moved.
there's still a bug when moving cursor horizontally inside a long line, in
that it's not detected that the screen needs redraw.
i'll leave the fix to someone else as i'm done with this buggy editor.